Shrine of Innos
Specific gold amounts cannot be donated. Instead for every level gained, up to 50, a donation can be made for an attribute bonus. The amount of gold required depends on donation prayers made and hero level. The bonus is not tied to guilds in any way and the different attributes receive bonuses inline with each other (pseudo random). In addition the hero can pray without donating for full health or mana once per day.